Avoid sending several requests to prometheus
Kafka charms are submitting several prometheus requests, every
time a cluster-changed or prometheus-changed event happens.

That stacks up requests for scraping the same targets, which end
up being all rendered as target options with juju.

Instead of creating new requests on the relation for every -changed
event, store the request_id on the app-level so it can be retrieved
later by other leader units; and generate one single request that
is updated as new units come up or go away.